+++ apps/mobile/src/components/app-tabs.web.tsx
@@ -54,7 +54,7 @@ export function CustomTabList(props: TabListProps) {
 		<View {...props} style={styles.tabListContainer}>
 			<ThemedView type="backgroundElement" style={styles.innerContainer}>
 				<ThemedText type="smallBold" style={styles.brandText}>
-					Expo School OS
+					Expo Starter
 				</ThemedText>
 
 				{props.children}